--- a/net/net_test.go
+++ b/net/net_test.go
@@ -3,11 +3,42 @@ package net
 import (
 	"net"
 	"net/http"
+	"net/netip"
 	"reflect"
 	"strings"
 	"testing"
 )
 
+func TestRemoteAddr(t *testing.T) {
+	for _, tt := range []struct {
+		name   string
+		input  string
+		want   netip.Addr
+		fwdHdr string
+	}{
+		{"no header1", "127.0.0.1", netip.MustParseAddr("127.0.0.1"), ""},
+		{"no header2", "1.2.3.4", netip.MustParseAddr("1.2.3.4"), ""},
+		{"no header3", "100.200.300.400", netip.Addr{}, ""},
+		{"no header4", "127.0.0.1:8080", netip.MustParseAddr("127.0.0.1"), ""},
+		{"single header1", "127.0.0.1", netip.MustParseAddr("172.16.0.1"), "172.16.0.1"},
+		{"invalid header", "127.0.0.1", netip.MustParseAddr("127.0.0.1"), "invalid header"},
+		{"multiple header1", "127.0.0.1", netip.MustParseAddr("172.16.0.1"), "172.16.0.1, 1.2.3.4, 8.7.6.5"}, // X-Forwarded-For with proxies in it
+		{"no header5", "2001:4860:0:2001::68", netip.MustParseAddr("2001:4860:0:2001::68"), ""},
+		{"single header2", "127.0.0.1", netip.MustParseAddr("2001:4860:0:2001::68"), "2001:4860:0:2001::68"},
+	} {
+		t.Run(tt.name, func(t *testing.T) {
+			r := &http.Request{RemoteAddr: tt.input, Header: make(http.Header)}
+			if tt.fwdHdr != "" {
+				r.Header.Set("x-forwarded-for", tt.fwdHdr)
+			}
+			got := RemoteAddr(r)
+
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("Unexpected IP address '%v'. Wanted '%v", got, tt.want)
+			}
+		})
+	}
+}
 func TestRemoteHost(t *testing.T) {
 	for _, tt := range []struct {
 		name   string
@@ -44,6 +75,37 @@ func BenchmarkRemoteHost(b *testing.B) {
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		RemoteHost(r)
+	}
+}
+
+func TestRemoteAddrFromLast(t *testing.T) {
+	for _, tt := range []struct {
+		name   string
+		input  string
+		want   netip.Addr
+		fwdHdr []string
+	}{
+		{"no header", "127.0.0.1", netip.MustParseAddr("127.0.0.1"), []string{}},
+		{"no header2", "1.2.3.4", netip.MustParseAddr("1.2.3.4"), []string{}},
+		{"no header3", "100.200.300.400", netip.Addr{}, []string{}},
+		{"no header4", "127.0.0.1:8080", netip.MustParseAddr("127.0.0.1"), []string{}},
+		{"single header", "127.0.0.1", netip.MustParseAddr("172.16.0.1"), []string{"172.16.0.1"}},
+		{"invalid  header", "127.0.0.1", netip.MustParseAddr("127.0.0.1"), []string{"invalid header"}},
+		{"invalid and remoteIp", "invalid-ip", netip.Addr{}, []string{"invalid, header"}},
+		{"multiple entries", "127.0.0.1", netip.MustParseAddr("8.7.6.5"), []string{"172.16.0.1", "1.2.3.4", "8.7.6.5"}},
+		{"2 entries", "127.0.0.1", netip.MustParseAddr("8.7.6.5"), []string{"1.2.3.4", "8.7.6.5"}},
+		{"single header2", "127.0.0.1", netip.MustParseAddr("8.7.6.5"), []string{"8.7.6.5"}},
+	} {
+		t.Run(tt.name, func(t *testing.T) {
+			r := &http.Request{RemoteAddr: tt.input, Header: make(http.Header)}
+			s := strings.Join(tt.fwdHdr, ", ")
+			r.Header.Set("x-forwarded-for", s)
+			got := RemoteAddrFromLast(r)
+
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("Unexpected IP address '%v'. Wanted '%v", got, tt.want)
+			}
+		})
 	}
 }
 
@@ -85,10 +147,38 @@ func BenchmarkRemoteHostFromLast(b *testing.B) {
 	}
 }
 
+func TestParseIPCIDRs(t *testing.T) {
+	for _, tt := range []struct {
+		input   []string
+		wantErr bool
+	}{
+		{[]string{"1.2.3.4.5"}, true},
+		{[]string{"1.2.3.4/"}, true},
+		{[]string{"1.2.3.4/245"}, true},
+		{[]string{"whatever"}, true},
+		{[]string{"1.2.3.4/24", "whatever"}, true},
+		{[]string{"1.2.3.4"}, false},
+		{[]string{"1.2.3.4/16"}, false},
+		{[]string{"1.2.3.4", "foo", "1.2.3.4/16"}, true},
+	} {
+		t.Run(strings.Join(tt.input, ","), func(t *testing.T) {
+			_, err := ParseIPCIDRs(tt.input)
+			if tt.wantErr && err == nil {
+				t.Logf("ParseIPCIDRs: %v", tt.input)
+				t.Errorf("parse error expected")
+			}
+			if !tt.wantErr && err != nil {
+				t.Errorf("parse error unexpected: %v", err)
+			}
+		})
+	}
+}
+
 func TestIPNetsParse(t *testing.T) {
 	for _, tt := range []struct {
 		input []string
 	}{
+		{[]string{"1.2.3.4.5"}},
 		{[]string{"1.2.3.4/"}},
 		{[]string{"1.2.3.4/245"}},
 		{[]string{"whatever"}},