Optimize: ClientIP(), Source(), SourceFromLast() predicates for a lot of args, benchmarks ipv6, too
```
% benchstat old.txt new.txt                                                                                                                                 optimize/source-predicates
name                  old time/op    new time/op    delta
ClientIP10-4             278ns ± 4%     107ns ± 5%   -61.45%  (p=0.000 n=7+10)
ClientIP100-4           2.45µs ±22%    0.14µs ±14%   -94.25%  (p=0.000 n=10+9)
ClientIP1k-4            20.8µs ± 3%     0.2µs ± 5%   -99.21%  (p=0.000 n=8+9)
ClientIP10k-4            207µs ± 2%       0µs ± 2%   -99.91%  (p=0.000 n=10+8)
ClientIP100k-4          2.23ms ± 2%    0.00ms ±10%   -99.99%  (p=0.000 n=9+10)
Source10-4               410ns ± 7%     237ns ±10%   -42.34%  (p=0.000 n=10+9)
Source100-4             2.20µs ± 6%    0.25µs ± 1%   -88.40%  (p=0.000 n=10+8)
Source1k-4              21.5µs ± 8%     0.3µs ±24%   -98.62%  (p=0.000 n=9+9)
Source10k-4              215µs ± 9%       0µs ± 8%   -99.85%  (p=0.000 n=9+9)
Source100k-4            2.24ms ± 1%    0.00ms ± 2%   -99.98%  (p=0.000 n=8+10)
SourceFromLast10-4       513ns ± 2%     141ns ± 4%   -72.60%  (p=0.000 n=9+10)
SourceFromLast100-4     2.32µs ± 6%    0.18µs ±18%   -92.44%  (p=0.000 n=10+9)
SourceFromLast1k-4      21.8µs ±11%     0.2µs ±14%   -99.06%  (p=0.000 n=9+10)
SourceFromLast10k-4      209µs ± 4%       0µs ± 3%   -99.89%  (p=0.000 n=10+9)
SourceFromLast100k-4    2.25ms ± 3%    0.00ms ± 4%   -99.99%  (p=0.000 n=9+9)

name                  old alloc/op   new alloc/op   delta
ClientIP10-4             16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
ClientIP100-4            16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
ClientIP1k-4             16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
ClientIP10k-4            16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
ClientIP100k-4           16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
Source10-4               48.0B ± 0%     32.0B ± 0%   -33.33%  (p=0.000 n=10+10)
Source100-4              48.0B ± 0%     32.0B ± 0%   -33.33%  (p=0.000 n=10+10)
Source1k-4               48.0B ± 0%     32.0B ± 0%   -33.33%  (p=0.000 n=10+10)
Source10k-4              48.0B ± 0%     32.0B ± 0%   -33.33%  (p=0.000 n=10+10)
Source100k-4             48.0B ± 0%     32.0B ± 0%   -33.33%  (p=0.000 n=10+10)
SourceFromLast10-4       80.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
SourceFromLast100-4      80.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
SourceFromLast1k-4       80.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
SourceFromLast10k-4      80.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
SourceFromLast100k-4     80.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)

name                  old allocs/op  new allocs/op  delta
ClientIP10-4              1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
ClientIP100-4             1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
ClientIP1k-4              1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
ClientIP10k-4             1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
ClientIP100k-4            1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
Source10-4                2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=10+10)
Source100-4               2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=10+10)
Source1k-4                2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=10+10)
Source10k-4               2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=10+10)
Source100k-4              2.00 ± 0%      1.00 ± 0%   -50.00%  (p=0.000 n=10+10)
SourceFromLast10-4        3.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
SourceFromLast100-4       3.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
SourceFromLast1k-4        3.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
SourceFromLast10k-4       3.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
SourceFromLast100k-4      3.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
```