@AlexanderYastrebov 2023-01-10
> Optimize: ClientIP(), Source(), SourceFromLast() predicates

Would it be possible to compare benchmarks before and after the change via https://pkg.go.dev/golang.org/x/perf/cmd/benchstat ?