@szuecs 2023-01-10
Yes maybe but I wouldn't care much or do you see any big advantage?