@AlexanderYastrebov 2023-01-25
If I read it correct benchmark shows increase of runtime and allocations for the most common usecase - predicates with a few IP addresses.

For large address lists I think it would not be practical to use existing predicates but rather create a new predicate(s) that could load subnets in bulk from some kind of data storage and utilize `netipx.IPSet`.