@szuecs 2023-01-25
@AlexanderYastrebov you miss the point of IPv6. Before it was not supported as far as I can tell and now it needs more space because of that so more allocations.