@AlexanderYastrebov 2023-01-25
> you miss the point of IPv6. Before it was not supported as far as I can tell

https://pkg.go.dev/net#ParseIP supports IPv6