@AlexanderYastrebov 2023-01-25
According to https://pkg.go.dev/net/netip
> Compared to the net.IP type, this package's Addr type takes less memory, is immutable, and is comparable (supports == and being a map key).

so I am wondering where the increase come from, maybe from `netipx.IPSet`...