@szuecs 2023-01-25
@AlexanderYastrebov it's because in IPv4 case it's only 4 bytes + slice header and Addr has 128 bit type + *Value as far as I understand, but they say different https://cs.opensource.google/go/go/+/refs/tags/go1.19.5:src/net/netip/netip.go;l=38