@AlexanderYastrebov 2023-01-25
OTOH absolute value increase is not dramatic as percentages because they are tiny.